requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
tensorflow==2.18.0
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
sqlalchemy==2.0.36
httpx==0.28.1
pdfplumber==0.11.4
pymupdf==1.25.3
pytesseract==0.3.13
Pillow==11.1.0
pdf2image==1.17.0
spacy==3.8.3
transformers==4.47.1
torch==2.5.1
regex==2024.11.6
langdetect==1.0.9
opencv-python==4.10.0.84
python-dateutil==2.9.0.post0
joblib==1.4.2
